+++ src/SettlerAbstract.sol
@@ -19,5 +19,5 @@ abstract contract SettlerAbstract is Permit2PaymentAbstract {
 
     function _hasMetaTxn() internal pure virtual returns (bool);
 
-    function _dispatch(uint256 i, bytes4 action, bytes calldata data) internal virtual returns (bool);
+    function _dispatch(uint256 i, uint256 action, bytes calldata data) internal virtual returns (bool);
 }