+++ src/chains/Mantle.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.25;
+pragma solidity =0.8.25;
 
 import {SettlerBase} from "../SettlerBase.sol";
 import {Settler} from "../Settler.sol";
@@ -29,7 +29,7 @@ abstract contract MantleMixin is FreeMemory, SettlerBase, DodoV2 {
         assert(block.chainid == 5000 || block.chainid == 31337);
     }
 
-    function _dispatch(uint256 i, bytes4 action, bytes calldata data)
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
         internal
         virtual
         override(SettlerBase, SettlerAbstract)
@@ -69,7 +69,7 @@ abstract contract MantleMixin is FreeMemory, SettlerBase, DodoV2 {
 contract MantleSettler is Settler, MantleMixin {
     constructor(bytes20 gitCommit) Settler(gitCommit) {}
 
-    function _dispatchVIP(bytes4 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
+    function _dispatchVIP(uint256 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
         return super._dispatchVIP(action, data);
     }
 
@@ -83,7 +83,7 @@ contract MantleSettler is Settler, MantleMixin {
         return super._isRestrictedTarget(target);
     }
 
-    function _dispatch(uint256 i, bytes4 action, bytes calldata data)
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
         internal
         override(SettlerAbstract, SettlerBase, MantleMixin)
         returns (bool)
@@ -100,7 +100,7 @@ contract MantleSettler is Settler, MantleMixin {
 contract MantleSettlerMetaTxn is SettlerMetaTxn, MantleMixin {
     constructor(bytes20 gitCommit) SettlerMetaTxn(gitCommit) {}
 
-    function _dispatchVIP(bytes4 action, bytes calldata data, bytes calldata sig)
+    function _dispatchVIP(uint256 action, bytes calldata data, bytes calldata sig)
         internal
         override
         DANGEROUS_freeMemory
@@ -110,7 +110,7 @@ contract MantleSettlerMetaTxn is SettlerMetaTxn, MantleMixin {
     }
 
     // Solidity inheritance is stupid
-    function _dispatch(uint256 i, bytes4 action, bytes calldata data)
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
         internal
         override(SettlerAbstract, SettlerBase, MantleMixin)
         returns (bool)